Separated VirtualizedTree for clarity
